Move abstract_udf to abstract subdirectory
